+++ src/main/java/org/sbrubbles/conditio/Handler.java
@@ -4,17 +4,17 @@ import java.util.function.BiFunction;
 import java.util.function.Predicate;
 
 /**
- * Handles conditions, producing the result to be returned by {@link Scope#signal(Condition) signal}. It may compute
- * this result by itself, or by {@linkplain Scope#restart(Restart.Option) delegating the work} to a restart of its
- * choice.
+ * Handles conditions, producing the result to be returned by {@link Scope#signal(Condition, Restart...) signal}.
+ * It may compute this result by itself, or by {@linkplain Scope#restart(Restart.Option) delegating the work} to a
+ * restart of its choice.
  * <p>
- * A handler doesn't necessarily <em>need</em> to handle a condition; it can decline to do so by returning
+ * A handler doesn't <em>need</em> to handle a condition; it can decline to do so by returning
  * {@link Handler#SKIP}. The signalling machinery will then search for another handler, bound later in the stack.
  * <p>
  * A handler can do two things:
  * <ul>
- *   <li>check if it can handle a given condition (with {@link #test(Object)}); and</li>
- *   <li>given a condition and its scope of origin, return a result (with {@link #apply(Object, Object)}).</li>
+ *   <li>check if it can handle a given condition (with {@link #test(Object) test}); and</li>
+ *   <li>given a condition and its scope of origin, return a result (with {@link #apply(Object, Object) apply}).</li>
  * </ul>
  * <p>
  * Since a handler works both as a {@linkplain Predicate<Object> predicate} and as a
@@ -29,10 +29,5 @@ public interface Handler extends Predicate<Condition>, BiFunction<Condition, Sco
    * Returned when a handler, for whatever reason, opts to not handle a particular condition. By returning this, other
    * handlers, bound later in the stack, will have the chance instead.
    */
-  Object SKIP = new Object() {
-    @Override
-    public String toString() {
-      return "Handler.SKIP";
-    }
-  };
+  Object SKIP = new Object();
 }